merge from istio

+++ mixer/adapter/kubernetesenv/kubernetesenv.go
@@ -1,4 +1,4 @@
-// Copyright 2017 Istio Authors
+// Copyright Istio Authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -13,8 +13,8 @@
 // limitations under the License.
 
 // nolint: lll
-//go:generate $GOPATH/src/istio.io/istio/bin/mixer_codegen.sh -a mixer/adapter/kubernetesenv/config/config.proto -x "-n kubernetesenv"
-//go:generate $GOPATH/src/istio.io/istio/bin/mixer_codegen.sh -t mixer/adapter/kubernetesenv/template/template.proto
+//go:generate $REPO_ROOT/bin/mixer_codegen.sh -a mixer/adapter/kubernetesenv/config/config.proto -x "-n kubernetesenv"
+//go:generate $REPO_ROOT/bin/mixer_codegen.sh -t mixer/adapter/kubernetesenv/template/template.proto
 
 // Package kubernetesenv provides functionality to adapt mixer behavior to the
 // kubernetes environment. Primarily, it is used to generate values as part
@@ -28,21 +28,24 @@ import (
 	"errors"
 	"fmt"
 	"net"
-	"os"
 	"strings"
 	"sync"
 	"time"
 
 	v1 "k8s.io/api/core/v1"
+	"k8s.io/client-go/dynamic"
 	k8s "k8s.io/client-go/kubernetes"
+	kubemeta "k8s.io/client-go/metadata"
 	_ "k8s.io/client-go/plugin/pkg/client/auth" // needed for auth
 	"k8s.io/client-go/tools/cache"
 	"k8s.io/client-go/tools/clientcmd"
 
 	"istio.io/istio/mixer/adapter/kubernetesenv/config"
 	ktmpl "istio.io/istio/mixer/adapter/kubernetesenv/template"
+	"istio.io/istio/mixer/adapter/metadata"
 	"istio.io/istio/mixer/pkg/adapter"
 	"istio.io/istio/pkg/kube/secretcontroller"
+	"istio.io/pkg/env"
 )
 
 const (
@@ -54,6 +57,8 @@ const (
 	defaultRefreshPeriod = 5 * time.Minute
 
 	defaultClusterRegistriesNamespace = "istio-system"
+
+	mixerClusterControllerPrefix = "mixer-cluster-"
 )
 
 var (
@@ -72,7 +77,8 @@ type (
 		sync.Mutex
 		controllers map[string]cacheController
 
-		kubeHandler *handler
+		kubeHandler             *handler
+		multiClusterWatcherInit sync.Once
 	}
 
 	handler struct {
@@ -94,17 +100,10 @@ var _ ktmpl.HandlerBuilder = &builder{}
 
 // GetInfo returns the Info associated with this adapter implementation.
 func GetInfo() adapter.Info {
-	return adapter.Info{
-		Name:        "kubernetesenv",
-		Impl:        "istio.io/istio/mixer/adapter/kubernetesenv",
-		Description: "Provides platform specific functionality for the kubernetes environment",
-		SupportedTemplates: []string{
-			ktmpl.TemplateName,
-		},
-		DefaultConfig: conf,
-
-		NewBuilder: func() adapter.HandlerBuilder { return newBuilder(newKubernetesClient) },
-	}
+	singletonBuilder := newBuilder(newKubernetesClient)
+	info := metadata.GetInfo("kubernetesenv")
+	info.NewBuilder = func() adapter.HandlerBuilder { return singletonBuilder }
+	return info
 }
 
 func (b *builder) SetAdapterConfig(c adapter.Config) {
@@ -117,12 +116,14 @@ func (b *builder) Validate() (ce *adapter.ConfigErrors) {
 	return
 }
 
+var kubeConfigVar = env.RegisterStringVar("KUBECONFIG", "", "Path for a kubeconfig file.")
+
 func (b *builder) Build(ctx context.Context, env adapter.Env) (adapter.Handler, error) {
 	paramsProto := b.adapterConfig
 	var controller cacheController
 	var controllers = make(map[string]cacheController)
 
-	path, exists := os.LookupEnv("KUBECONFIG")
+	path, exists := kubeConfigVar.Lookup()
 	if !exists {
 		path = paramsProto.KubeconfigPath
 	}
@@ -133,7 +134,7 @@ func (b *builder) Build(ctx context.Context, env adapter.Env) (adapter.Handler, 
 	// provide this basic functionality before.
 	b.Lock()
 	defer b.Unlock()
-	_, found := b.controllers[path]
+	_, found := b.controllers[mixerClusterControllerPrefix+path]
 	if !found {
 		clientset, err := b.newClientFn(path, env)
 		if err != nil {
@@ -143,8 +144,8 @@ func (b *builder) Build(ctx context.Context, env adapter.Env) (adapter.Handler, 
 		if err != nil {
 			return nil, fmt.Errorf("could not create new cache controller: %v", err)
 		}
-		controllers[path] = controller
-		b.controllers[path] = controller
+		controllers[mixerClusterControllerPrefix+path] = controller
+		b.controllers[mixerClusterControllerPrefix+path] = controller
 	} else {
 		for clusterID := range b.controllers {
 			controllers[clusterID] = b.controllers[clusterID]
@@ -161,8 +162,16 @@ func (b *builder) Build(ctx context.Context, env adapter.Env) (adapter.Handler, 
 	b.kubeHandler = &kubeHandler
 
 	if !found {
-		if err := initMultiClusterSecretController(b, path, env); err != nil {
-			return nil, fmt.Errorf("could not create remote controllers: %v", err)
+		// only init a single watcher for multicluster secrets for a given Builder.
+		// if init fails, try again on next Build().
+		// This should prevent an ever-growing number of watchers for a Builder.
+		var initErr error
+		b.multiClusterWatcherInit.Do(func() {
+			initErr = initMultiClusterSecretController(b, path, env)
+		})
+		if initErr != nil {
+			b.multiClusterWatcherInit = sync.Once{}
+			return nil, fmt.Errorf("could not create remote controllers: %v", initErr)
 		}
 	}
 
@@ -191,9 +200,10 @@ func runNewController(b *builder, clientset k8s.Interface, env adapter.Env) (cac
 
 func newBuilder(clientFactory clientFactoryFn) *builder {
 	return &builder{
-		newClientFn:   clientFactory,
-		controllers:   make(map[string]cacheController),
-		adapterConfig: conf,
+		newClientFn:             clientFactory,
+		controllers:             make(map[string]cacheController),
+		adapterConfig:           conf,
+		multiClusterWatcherInit: sync.Once{},
 	}
 }
 
@@ -224,8 +234,10 @@ func (h *handler) GenerateKubernetesAttributes(ctx context.Context, inst *ktmpl.
 }
 
 func (h *handler) Close() error {
-	for clusterID := range h.builder.controllers {
-		h.builder.deleteCacheController(clusterID)
+	for clusterID := range h.k8sCache {
+		if !strings.HasPrefix(clusterID, mixerClusterControllerPrefix) {
+			_ = h.builder.deleteCacheController(clusterID)
+		}
 	}
 	h.builder.Lock()
 	h.builder.kubeHandler = nil
@@ -256,16 +268,15 @@ func (h *handler) findPod(uid string) (cacheController, *v1.Pod, bool) {
 	return c, pod, found
 }
 
+//The name of workload may contain '.'
 func keyFromUID(uid string) string {
 	if ip := net.ParseIP(uid); ip != nil {
 		return uid
 	}
 	fullname := strings.TrimPrefix(uid, kubePrefix)
-	if strings.Contains(fullname, ".") {
-		parts := strings.Split(fullname, ".")
-		if len(parts) == 2 {
-			return key(parts[1], parts[0])
-		}
+	dotPos := strings.LastIndex(fullname, ".")
+	if dotPos != -1 {
+		return key(fullname[dotPos+1:], fullname[:dotPos])
 	}
 	return fullname
 }
@@ -361,7 +372,7 @@ func newKubernetesClient(kubeconfigPath string, env adapter.Env) (k8s.Interface,
 	return k8s.NewForConfig(config)
 }
 
-func (b *builder) createCacheController(k8sInterface k8s.Interface, clusterID string) error {
+func (b *builder) createCacheController(k8sInterface k8s.Interface, _ kubemeta.Interface, _ dynamic.Interface, clusterID string) error {
 	controller, err := runNewController(b, k8sInterface, b.kubeHandler.env)
 	if err == nil {
 		b.Lock()
@@ -373,11 +384,17 @@ func (b *builder) createCacheController(k8sInterface k8s.Interface, clusterID st
 		b.kubeHandler.Unlock()
 
 		b.kubeHandler.env.Logger().Infof("created remote controller %s", clusterID)
-	} else {
-		b.kubeHandler.env.Logger().Errorf("error on creating remote controller %s err = %v", clusterID, err)
+		return nil
 	}
 
-	return err
+	return b.kubeHandler.env.Logger().Errorf("error on creating remote controller %s err = %v", clusterID, err)
+}
+
+func (b *builder) updateCacheController(k8sInterface k8s.Interface, _ kubemeta.Interface, _ dynamic.Interface, clusterID string) error {
+	if err := b.deleteCacheController(clusterID); err != nil {
+		return err
+	}
+	return b.createCacheController(k8sInterface, nil, nil, clusterID)
 }
 
 func (b *builder) deleteCacheController(clusterID string) error {
@@ -395,14 +412,14 @@ func (b *builder) deleteCacheController(clusterID string) error {
 	return nil
 }
 
+var clusterNsVar = env.RegisterStringVar("POD_NAMESPACE", defaultClusterRegistriesNamespace, "Namespace for the Mixer pod (Downward API).")
+
 func initMultiClusterSecretController(b *builder, kubeconfig string, env adapter.Env) (err error) {
 	var clusterNs string
 
 	paramsProto := b.adapterConfig
 	if clusterNs = paramsProto.ClusterRegistriesNamespace; clusterNs == "" {
-		if clusterNs = os.Getenv("POD_NAMESPACE"); clusterNs == "" {
-			clusterNs = defaultClusterRegistriesNamespace
-		}
+		clusterNs = clusterNsVar.Get()
 	}
 
 	kubeClient, err := b.newClientFn(kubeconfig, env)
@@ -410,10 +427,8 @@ func initMultiClusterSecretController(b *builder, kubeconfig string, env adapter
 		return fmt.Errorf("could not create K8s client: %v", err)
 	}
 
-	err = secretcontroller.StartSecretController(kubeClient, b.createCacheController, b.deleteCacheController, clusterNs)
-	if err != nil {
-		return fmt.Errorf("could not start secret controller: %v", err)
-	}
+	_ = secretcontroller.StartSecretController(kubeClient, b.createCacheController,
+		b.updateCacheController, b.deleteCacheController, clusterNs)
 
 	return nil
 }
